+++ Content.Shared/Cargo/BUI/CargoConsoleInterfaceState.cs
@@ -8,10 +8,10 @@ public sealed class CargoConsoleInterfaceState : BoundUserInterfaceState
     public string Name;
     public int Count;
     public int Capacity;
-    public int Balance;
+    public ulong Balance;
     public List<CargoOrderData> Orders;
 
-    public CargoConsoleInterfaceState(string name, int count, int capacity, int balance, List<CargoOrderData> orders)
+    public CargoConsoleInterfaceState(string name, int count, int capacity, ulong balance, List<CargoOrderData> orders)
     {
         Name = name;
         Count = count;
